refactor and fix skillmap/teachertool
