+++ webapp/src/browserworkspace.ts
@@ -1,13 +1,12 @@
 import * as db from "./db";
 import * as workspace from "./workspace";
-import * as pxteditor from "../../pxteditor";
 
 let headerDb: db.Table;
 let textDb: db.Table;
 
 type Header = pxt.workspace.Header;
-type ScriptText = pxteditor.workspace.ScriptText;
-type WorkspaceProvider = pxteditor.workspace.WorkspaceProvider;
+type ScriptText = pxt.workspace.ScriptText;
+type WorkspaceProvider = pxt.workspace.WorkspaceProvider;
 
 type TextDbEntry = {
     files?: ScriptText,
@@ -86,7 +85,7 @@ async function listAsync(): Promise<pxt.workspace.Header[]> {
     return headerDb.getAllAsync() as Promise<Header[]>;
 }
 
-async function getAsync(h: Header): Promise<pxteditor.workspace.File> {
+async function getAsync(h: Header): Promise<pxt.workspace.File> {
     const hdrProm = headerDb.getAsync(h.id)
     const textProm = textDb.getAsync(h.id)
     let [hdrResp, textResp] = await Promise.all([hdrProm, textProm]) as [Header, TextDbEntry]
@@ -122,7 +121,7 @@ function setCoreAsync(headers: db.Table, texts: db.Table, h: Header, prevVer: an
     return headerRes
 }
 
-export async function copyProjectToLegacyEditor(header: Header, script: pxteditor.workspace.ScriptText, majorVersion: number): Promise<void> {
+export async function copyProjectToLegacyEditor(header: Header, script: pxt.workspace.ScriptText, majorVersion: number): Promise<void> {
     const prefix = pxt.appTarget.appTheme.browserDbPrefixes && pxt.appTarget.appTheme.browserDbPrefixes[majorVersion];
 
     const oldHeaders = new db.Table(prefix ? `${prefix}-header` : `header`);
