+++ webapp/src/idbworkspace.ts
@@ -1,8 +1,6 @@
-import * as pxteditor from "../../pxteditor";
-
 type Header = pxt.workspace.Header;
-type ScriptText = pxteditor.workspace.ScriptText;
-type WorkspaceProvider = pxteditor.workspace.WorkspaceProvider;
+type ScriptText = pxt.workspace.ScriptText;
+type WorkspaceProvider = pxt.workspace.WorkspaceProvider;
 
 interface StoredText {
     id: string;
@@ -251,7 +249,7 @@ async function listAsync(): Promise<pxt.workspace.Header[]> {
     return db.getAllAsync<pxt.workspace.Header>(HEADERS_TABLE);
 }
 
-async function getAsync(h: Header): Promise<pxteditor.workspace.File> {
+async function getAsync(h: Header): Promise<pxt.workspace.File> {
     const db = await getCurrentDbAsync();
     const res = await db.getAsync<StoredText>(TEXTS_TABLE, h.id);
     return {
@@ -324,7 +322,7 @@ export async function getObjectStoreAsync<T>(storeName: string) {
     return db.getObjectStoreWrapper<T>(storeName);
 }
 
-export async function copyProjectToLegacyEditorAsync(header: Header, script: pxteditor.workspace.ScriptText, majorVersion: number): Promise<void> {
+export async function copyProjectToLegacyEditorAsync(header: Header, script: pxt.workspace.ScriptText, majorVersion: number): Promise<void> {
     const prefix = pxt.appTarget.appTheme.browserDbPrefixes && pxt.appTarget.appTheme.browserDbPrefixes[majorVersion];
 
     const oldDB = await getDbAsync(prefix);
